identify who is login accross the app

+++ app/Http/Controllers/AdminController.php
@@ -21,137 +21,109 @@ class AdminController extends Controller
 
     public function index()
     {
-        $users = User::with('profile')->orderByDesc('created_at')->get();
-        if ($users) {
-            return Inertia::render('Backend/User', [
-                'users' => $users
-            ]);
+        if (auth()->user()){
+            $users = User::with('profile')->orderByDesc('created_at')->get();
+            if ($users) {
+                return Inertia::render('Backend/User', [
+                    'users' => $users
+                ]);
 
-        } else {
-            return redirect()->route('user.dashboard')->with('success', 'Success, your store have been updated.');
-
+            }
+        }else{
+            return redirect()->route('home')->with('message', 'error, Not Authenticated.');
         }
+
+
     }
 
 
     public function store(Request $request)
     {
 
-                $request->validate([
-                    'username' => 'required|string|min:3,max:10|unique:users,username',
-                    'password' => 'required',
-                    'email' => 'required|unique:users,email',
-                    'contact' => 'required',
+
+//        dd($request);
+        if (auth()->user()){
+            $request->validate([
+                'username' => 'required|string|min:3,max:10|unique:users,username',
+                'password' => 'required',
+                'email' => 'required|unique:users,email',
+                'contact' => 'required',
+            ]);
+
+            DB::transaction(function () use ($request) {
+
+                $data['user'] = User::create([
+                    'username' => $request->username,
+                    'email' => $request->email,
+                    'contact' => $request->contact,
+                    'password' => Hash::make($request->password),
                 ]);
 
-                    DB::transaction(function () use ($request) {
+                $data['profile'] =  Profile::create([
+                    'user_id' => $data['user']->id,
 
-                        $data['user'] = User::create([
-                            'username' => $request->username,
-                            'email' => $request->email,
-                            'contact' => $request->contact,
-                            'password' => Hash::make($request->password),
-                        ]);
+                ]);
 
-                        $data['profile'] =  Profile::create([
-                            'user_id' => $data['user']->id,
+                return redirect()->route('user.dashboard')->with('message', 'Success, User Created Successfully.');
+            });
 
-                        ]);
-
-                    });
+        }else{
+            return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
+        }
 
 
         }
 
 
-    public function updateUserRole(User $user)
-    {
-        $new_role = Role::whereNotIn('id', $user->id)->first();
-        $user->update();
-        if ($user) {
-            return response()->json([
-                'message' => "User role  updated ",
-                'data' => $user,
-                'code' => 200
-            ],200);
-        } else {
-            return response()->json([
-                'message' => "Request  unsuccessful ",
-                'code'=>400
-            ],400);
-        }
-    }
 
     public function getUpdateForm(User $user)
     {
+if (auth()->user()){
+    $user = User::findOrFail($user->id);
+    return Inertia::render('Backend/EditUser', [
+        'id' => $user->id,
+        'username' => $user->username,
+        'email' => $user->email,
+        'contact' => $user->contact
 
-        $user = User::findOrFail($user->id);
-        return Inertia::render('Backend/EditUser', [
-            'id' => $user->id,
-            'username' => $user->username,
-            'email' => $user->email,
-            'contact' => $user->contact
-
-        ]);
+    ]);
+}
+    else{
+        return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
+    }
 
     }
 
 
     public function update(Request $request, User $user)
     {
-        $this->validate($request, [
-            'username' => 'required|max:35',
-            'email' => 'required|email|unique:users,email,'.$user->id,
-            'contact' => 'required'
-        ]);
+        if (auth()->user()){
+            $this->validate($request, [
+                'username' => 'required|max:35',
+                'email' => 'required|email|unique:users,email,'.$user->id,
+                'contact' => 'required'
+            ]);
 
-        User::findOrFail($user->id)->update([
-            'username' => $request->username,
-            'email' => $request->email,
-            'contact' => $request->contact,
-        ]);
-        return redirect()->route('dashboard.users');
+            User::findOrFail($user->id)->update([
+                'username' => $request->username,
+                'email' => $request->email,
+                'contact' => $request->contact,
+            ]);
 
-    }
-    public function AdminUpdateUserProfile(Request $request, User $user)
-    {
-        $request->validate([
-            'first_name' => 'nullable|string',
-            'last_name' => 'nullable|string',
-            'username' => 'required|string|min:3,max:10|unique:users,username',
-            'email' => ['required|email'],
-            'contact' => ['required|number']
-        ]);
+            return redirect()->route('user.dashboard')->with('message', 'Success, User Updated Successfully.');
 
-        $data['user'] = $user->update([
-            'username' => $request->username,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-        ]);
-        $user->profile()->update([
-            'first_name' => $request->first_name,
-            'last_name' => $request->last_name,
-            'image' =>  Null,
-        ]);
-        $data['user'] = User::with('profile')->whereId($user->id)->first();
-        if ($data) {
-            return response()->json([
-                'message' => "User  updated ",
-                'data' => $data,
-                'code' => 200
-            ],200);
-        } else {
-            return response()->json([
-                'message' => "Request  unsuccessful ",
-                'code'=>400
-            ],400);
+        }else{
+            return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
         }
+
+
     }
+
 
     public function userDetails(User $user)
     {
-
-        $user = User::with('profile')->whereId($user->id)->first();
+        if (auth()->user()){
+            $user = User::with('profile')->whereId($user->id)->first();
 
             return Inertia::render('Backend/ViewUser', [
                 'id' => $user->id,
@@ -162,6 +134,10 @@ class AdminController extends Controller
                 'last_name' => $user->profile->last_name,
                 'photo' => $user->profile->image,
             ]);
+        }else{
+            return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
+        }
+
 
 
 
@@ -169,27 +145,36 @@ class AdminController extends Controller
 
     public function getDeleteForm(User $user)
     {
+        if (auth()->user()){
+            $user = User::findOrFail($user->id);
+            return Inertia::render('Backend/DeleteUser', [
+                'id' => $user->id,
+                'username' => $user->username,
+                'email' => $user->email,
+                'contact' => $user->contact
+            ]);
+        }else{
+            return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
+        }
 
-        $user = User::findOrFail($user->id);
-        return Inertia::render('Backend/DeleteUser', [
-            'id' => $user->id,
-            'username' => $user->username,
-            'email' => $user->email,
-            'contact' => $user->contact
-        ]);
 
     }
     public function destroy(User $user)
     {
-        $user = User::findOrFail($user->id);
+        if (auth()->user()){
+            $user = User::findOrFail($user->id);
 
-        $user->delete();
-        return redirect()->route('dashboard.users');
+            $user->delete();
+            return redirect()->route('user.dashboard')->with('message', 'Success, User Deleted Successfully.');
+        }else{
+            return redirect()->route('home')->with('error', 'fail, Not Authenticated.');
+        }
+
     }
 
     public function login(Request $request)
     {
-
+//   dd($request);
         $credentials = $request->validate([
             'username' => 'required|string',
             'password' => ['required'],
@@ -214,10 +199,7 @@ class AdminController extends Controller
             // Load user from database
             if ($user && !Hash::check($request->password, $user->password)) {
 
-                return response()->json([
-                    "message" => 'The Provided password credentials are not correct',
-                    "code" => 403,
-                ], 403);
+                return redirect()->route('user.login')->with('error', 'Fail, Password Incorrect.');
             }
 
             if (Hash::check($request->password, $user->password)) {
@@ -227,25 +209,15 @@ class AdminController extends Controller
 
                 if (!Auth::attempt($credentials, $remember)) {
 
-                    return response()->json([
-                        'messages' => 'The Provided credentials are not correct',
-                        "code" => 403
-                    ], 403); //bad request or invalide
+                    return redirect()->route('user.login')->with('error', 'Fail, Password Incorrect.');
                 }
                 // todo to be uncommented soon for device authentication still
 
                 $user = Auth::user();
-                $token = $user->createToken('main')->plainTextToken;
 
-                $user = User::whereId($user->id)->with('profile')->first();
-                return Redirect::route('user.dashboard');
+                return redirect()->route('user.dashboard')->with('message', 'Success, User Login Successfully.');
             }
-        } else {
 
-            return response([
-                'message' => 'User Name Incorrect',
-                'code' => 409
-            ],409);
         }
     }
 
@@ -253,43 +225,16 @@ class AdminController extends Controller
     public function logout()
     {
 
-        if (request()->user('sanctum')) {
+        if (auth()->id()) {
             $user = Auth::user();
-            // Revoke the token that was used to authenticate the current request...
-            $user->currentAccessToken()->delete();
+
+            Auth::logout();
+            request()->session()->invalidate();
             // response to this message if successful logout
-            return redirect()->route('home');
+            return redirect()->route('home')->with('message', 'Success, User Logout Successfully.');
         }
     }
 
-    // Auto complete on user name
 
-    public function search(Request $request)
-    {
-
-        if (request()->user('sanctum')) {
-            $data['users'] = User::where('username', 'LIKE', "%{$request->username}%")->get();
-
-            //checking if request exit
-            if ($data) {
-                return response()->json([
-                    'data' => $data,
-                    'message' => "Record found",
-                    'code' => 200,
-                ], 200);
-            } else {
-                return response()->json([
-                    'message' => "No record found",
-                    'code' => 404,
-                ], 404);
-            }
-        } else {
-            return response()->json([
-                "message" => "Please Login First",
-                "code" => 403,
-
-            ], 403);
-        }
-    }
 
 }
